# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CleanOps Pro - Multi-tenant SaaS API for cleaning businesses
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /companies/me:
    get:
      operationId: getMyCompany
      tags: [companies]
      summary: Get current user's company
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

    put:
      operationId: updateMyCompany
      tags: [companies]
      summary: Update current user's company
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyRequest"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientListResponse"

    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientDetail"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Client deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [owner, admin, office, technician, super_admin]
        company_id:
          type: integer
        avatar_url:
          type: string
          nullable: true
      required:
        - id
        - email
        - first_name
        - last_name
        - role
        - company_id
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logo_url:
          type: string
          nullable: true
        subscription_status:
          type: string
          enum: [active, past_due, canceled, trialing]
        plan:
          type: string
          enum: [starter, growth, enterprise]
        employee_count:
          type: integer
        pay_cadence:
          type: string
          enum: [weekly, biweekly, semimonthly]
        created_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - subscription_status
        - plan
        - employee_count
        - pay_cadence
        - created_at
    UpdateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        logo_url:
          type: string
          nullable: true
        pay_cadence:
          type: string
          enum: [weekly, biweekly, semimonthly]
        geo_fence_threshold_ft:
          type: integer
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [owner, admin, office, technician]
        pay_rate:
          type: number
          nullable: true
        pay_type:
          type: string
          enum: [hourly, per_job, fee_split]
          nullable: true
        is_active:
          type: boolean
        hire_date:
          type: string
          format: date
          nullable: true
        avatar_url:
          type: string
          nullable: true
        productivity_pct:
          type: number
          nullable: true
      required:
        - id
        - email
        - first_name
        - last_name
        - role
        - is_active
...
    Client:
      type: object
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        loyalty_points:
          type: integer
        created_at:
          type: string
          format: date-time
      required:
        - id
        - first_name
        - last_name
        - loyalty_points
        - created_at
    ClientDetail:
      allOf:
        - $ref: "#/components/schemas/Client"
        - type: object
          properties:
            notes:
              type: string
              nullable: true
            lat:
              type: number
              nullable: true
            lng:
              type: number
              nullable: true
            recent_jobs:
              type: array
              items:
                $ref: "#/components/schemas/JobSummary"
            total_jobs:
              type: integer
            total_revenue:
              type: number
            last_service_date:
              type: string
              format: date
              nullable: true
            assigned_cleaner:
              $ref: "#/components/schemas/User"
              nullable: true
    ClientListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
    CreateClientRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - first_name
        - last_name
    UpdateClientRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        zip:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    JobSummary:
      type: object
      properties:
        id:
          type: integer
        service_type:
          type: string
        status:
          type: string
        scheduled_date:
          type: string
          format: date
        client_name:
          type: string
        base_fee:
          type: number
      required:
        - id
        - service_type
        - status
        - scheduled_date
        - client_name
        - base_fee
...