# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CleanOps Pro - Multi-tenant SaaS API for cleaning businesses
 * OpenAPI spec version: 1.0.0
 */

export interface SuccessResponse {
  success: boolean;
  message?: string;
}